twitter_google_drive_app/requirements.txt
Flask
Flask-Dance
requests
tweepy
pandas
numpy
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
celery
redis
pytest
pytest-mock
gunicorn
